+++ torch_skeleton/datasets/__init__.py
@@ -1,2 +1,4 @@
-from .ntu import NTUDataset
+from .ntu import NTU
 from .nw_ucla import UCLA
+
+from .base_dataset import DiskCache
