--- torch_skeleton/datasets/ntu/__init__.py
@@ -1,4 +1,0 @@
-from .ntu import *
-from .utils import *
-
-from .dataset import *
